expert/requirements.txt
fastapi==0.99.1
python-jose[cryptography]
passlib
psycopg2-binary
langchain
chromadb
openai
pysqlite3-binary 
tiktoken
